Change visibility of WithdrawalService and stub; refactor packages
