+++ src/main/java/money/boku/Main.java
@@ -3,6 +3,10 @@ package money.boku;
 import io.javalin.Javalin;
 import io.javalin.http.HttpStatus;
 import money.boku.account.*;
+import money.boku.operations.*;
+import money.boku.operations.withdrawal.WithdrawalNotFoundException;
+import money.boku.operations.withdrawal.WithdrawalRequestWatcher;
+import money.boku.operations.withdrawal.WithdrawalServiceStub;
 
 /**
  * Starts up the server.
