+++ src/main/java/money/boku/operations/MoneyService.java
@@ -1,7 +1,9 @@
-package money.boku;
+package money.boku.operations;
 
 import money.boku.account.AccountNotFoundException;
 import money.boku.account.NegativeBalanceRequestException;
+import money.boku.operations.withdrawal.WithdrawalNotFoundException;
+import money.boku.operations.withdrawal.WithdrawalService;
 
 import java.math.BigDecimal;
 import java.util.UUID;
