+++ src/main/java/money/boku/operations/Transfer.java
@@ -1,4 +1,4 @@
-package money.boku;
+package money.boku.operations;
 
 import java.math.BigDecimal;
 import java.util.UUID;
@@ -6,5 +6,5 @@ import java.util.UUID;
 /**
  * Passed to API when requesting transfer from one user account to another.
  */
-public record WithdrawalRequest(UUID accountId, String withdrawalAddress, BigDecimal amount) {
+public record Transfer(UUID from, UUID to, BigDecimal amount) {
 }
