+++ src/main/java/money/boku/operations/withdrawal/WithdrawalRequestResponse.java
@@ -1,4 +1,4 @@
-package money.boku;
+package money.boku.operations.withdrawal;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
