+++ src/main/java/money/boku/operations/withdrawal/WithdrawalService.java
@@ -1,0 +1,34 @@
+package money.boku.operations.withdrawal;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+// Changed imports here as original did not compile.
+
+
+public interface WithdrawalService {
+    /**
+     * Request a withdrawal for given address and amount. Completes at random moment between 1 and 10 seconds
+     * @param id - a caller generated withdrawal id, used for idempotency
+     * @param address - an address withdraw to, can be any arbitrary string
+     * @param amount - an amount to withdraw (please replace T with type you want to use)
+     * @throws IllegalArgumentException in case there's different address or amount for given id
+     */
+    void requestWithdrawal(WithdrawalId id, Address address, BigDecimal amount); // Please substitute T with prefered type
+
+    /**
+     * Return current state of withdrawal
+     * @param id - a withdrawal id
+     * @return current state of withdrawal
+     * @throws IllegalArgumentException in case there no withdrawal for the given id
+     */
+    WithdrawalState getRequestState(WithdrawalId id);
+
+    enum WithdrawalState {
+        PROCESSING, COMPLETED, FAILED
+    }
+
+    record WithdrawalId(UUID value) {}
+    record Address(String value) {}
+}
+
