+++ src/main/java/money/boku/operations/withdrawal/WithdrawalServiceStub.java
@@ -1,44 +1,14 @@
-package money.boku;
+package money.boku.operations.withdrawal;
 
 import java.math.BigDecimal;
 import java.util.Objects;
-import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ThreadLocalRandom;
 
-// Changed imports here as original did not compile.
-import static money.boku.WithdrawalService.WithdrawalState.COMPLETED;
-import static money.boku.WithdrawalService.WithdrawalState.FAILED;
-import static money.boku.WithdrawalService.WithdrawalState.PROCESSING;
+import static money.boku.operations.withdrawal.WithdrawalService.WithdrawalState.*;
 
-interface WithdrawalService {
-    /**
-     * Request a withdrawal for given address and amount. Completes at random moment between 1 and 10 seconds
-     * @param id - a caller generated withdrawal id, used for idempotency
-     * @param address - an address withdraw to, can be any arbitrary string
-     * @param amount - an amount to withdraw (please replace T with type you want to use)
-     * @throws IllegalArgumentException in case there's different address or amount for given id
-     */
-    void requestWithdrawal(WithdrawalId id, Address address, BigDecimal amount); // Please substitute T with prefered type
-
-    /**
-     * Return current state of withdrawal
-     * @param id - a withdrawal id
-     * @return current state of withdrawal
-     * @throws IllegalArgumentException in case there no withdrawal for the given id
-     */
-    WithdrawalState getRequestState(WithdrawalId id);
-
-    enum WithdrawalState {
-        PROCESSING, COMPLETED, FAILED
-    }
-
-    record WithdrawalId(UUID value) {}
-    record Address(String value) {}
-}
-
-class WithdrawalServiceStub implements WithdrawalService {
+public class WithdrawalServiceStub implements WithdrawalService {
     private final ConcurrentMap<WithdrawalId, Withdrawal> requests = new ConcurrentHashMap<>();
 
     @Override
@@ -70,4 +40,3 @@ class WithdrawalServiceStub implements WithdrawalService {
         }
     }
 }
-
