+++ src/test/java/money/boku/MoneyAPIFunctionalTest.java
@@ -4,6 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.javalin.Javalin;
 import io.javalin.testtools.HttpClient;
 import io.javalin.testtools.JavalinTest;
+import money.boku.operations.withdrawal.WithdrawalRequestResponse;
+import money.boku.operations.withdrawal.WithdrawalService;
+import money.boku.operations.withdrawal.WithdrawalStateResponse;
 import okhttp3.Response;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
